+++ CHANGELOG.md
@@ -10,3 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - ChatGPT engine (ChatCompletion, Functions)
 - DALL-E2 engine
+
+## [0.2] - 2023-10-14
+
+### Added
+- JSON dumper for chats and settings
+- Added detailed readme